* YM Player now works (tested on linux)
git-svn-id: http://wiituka.googlecode.com/svn/trunk@32 1f7193ac-38c1-11de-8e0c-53a27eea117e

+++ source/ymlib/Ym2149Ex.h
@@ -91,7 +91,7 @@ public:
 		~CYm2149Ex();
 
 		void	reset(void);
-		void	update(signed short *pSampleBuffer,ymint nbSample);
+		void	update(int *pSampleBuffer,ymint nbSample);
 
 		void	setClock(ymu32 _clock);
 		void	writeRegister(ymint reg,ymint value);
